Develop

+++ src/main/java/com/nhnacademy/task/service/impl/CommentServiceImpl.java
@@ -19,9 +19,9 @@ import com.nhnacademy.task.repository.CommentRepository;
 import com.nhnacademy.task.repository.TaskRepository;
 import com.nhnacademy.task.service.CommentService;
 import java.util.List;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.RequestBody;
 
 @RequiredArgsConstructor
@@ -31,16 +31,15 @@ public class CommentServiceImpl implements CommentService {
     private final TaskRepository taskRepository;
 
     @Override
+    @Transactional
     public String registerComment(@RequestBody String commentContent, Long projectNum, Long taskNum,
                                   String writerId) {
-        Optional<Task> task = taskRepository.findById(taskNum);
+        Task task = taskRepository.findById(taskNum)
+                .orElseThrow(() -> new IllegalArgumentException("해당 task가 존재하지 않습니다."));
 
-        if (task.isEmpty()) {
-            return "해당 task가 존재하지 않습니다.";
-        }
         Comment comment = Comment.builder()
                 .commentContent(commentContent)
-                .task(task.get())
+                .task(task)
                 .writerId(writerId)
                 .build();
 
@@ -50,22 +49,21 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
+    @Transactional(readOnly = true)
     public List<CommentResponseDto> getAllComment(Long projectNum, Long taskNum) {
         Task task = taskRepository.findById(taskNum)
-                .orElseThrow(() -> new RuntimeException("해당 task가 존재하지 않습니다."));
+                .orElseThrow(() -> new IllegalArgumentException("해당 task가 존재하지 않습니다."));
 
         return commentRepository.findByTask(task);
     }
 
     @Override
+    @Transactional
     public String updateComment(String commentContent, Long projectNum, Long taskNum,
                                 Long commentNum) {
-        Optional<Comment> comment = commentRepository.findById(commentNum);
-        if (comment.isEmpty()) {
-            return "해당 comment가 존재하지 않습니다.";
-        }
+        Comment updateComment = commentRepository.findById(commentNum)
+                .orElseThrow(() -> new IllegalArgumentException("해당 comment가 존재하지 않습니다."));
 
-        Comment updateComment = comment.get();
         updateComment.setCommentContent(commentContent);
 
         commentRepository.save(updateComment);
@@ -74,7 +72,12 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
+    @Transactional
     public String deleteComment(Long projectNum, Long taskNum, Long commentNum) {
+        if (!commentRepository.existsById(commentNum)) {
+            return "해당 comment가 존재하지 않습니다.";
+        }
+
         commentRepository.deleteById(commentNum);
 
         return "comment가 삭제되었습니다.";
